+++ packages/stripe_js/lib/src/api/payment_intents/confirm_payment_options.dart
@@ -11,6 +11,7 @@ class ConfirmPaymentOptions with _$ConfirmPaymentOptions {
   const factory ConfirmPaymentOptions({
     /// The Elements instance that was used to create the Payment Element.
     @ElementsConverter() required Elements elements,
+    String? clientSecret,
 
     /// Parameters that will be passed on to the Stripe API.
     /// Refer to the Payment Intents API for a full list of parameters.